requirements.txt
langchain
langchain-community
langchain-openai
pymupdf
supabase
python-dotenv
openai
